+++ src/TwentyTwenty.DomainDriven/EventSourcing/IEventDescriptor.cs
@@ -5,11 +5,8 @@ namespace TwentyTwenty.DomainDriven.EventSourcing
     public interface IEventDescriptor
     {
         Guid Id { get; }
-
-        int Version { get; }
-
+        long Version { get; }
         DateTime TimeStamp { get; }
-
         IDomainEvent Data { get; }
     }
